IDC UT: Add mutex to guarantee behaviour of mock feed cache
We've been seeing IDC UT flakes that boil down to this:

- Test sets up a scenario where there will be an error, that should be reported
in GlobalThreatFeedStatus.

- Test checks for that error in GlobalThreatFeedStatus and finds no errors
there.

The only way I can make sense of the observed IDC UT flakes is that the program
is showing undefined behaviour because of data being written and read from
different goroutines without any synchronization.  Strictly speaking, the
golang spec says that behaviour is undefined in this case - i.e. I guess that the
compiler is within its rights to generate code that might behave unintuitively.

If we take a detailed look at what happens in the UT code, versus what happens
in production code:

- The production code (GlobalThreatFeedCacher) uses internal request and
response channels, and an additional goroutine, which adds up to ensuring
correctly synchronized behaviour when the GlobalThreatFeedCacher is called from
multiple calling goroutines.  It also uses `DeepCopy()` each time that it
returns an `apiv3.GlobalThreatFeed` pointer, which means that the callers get a
pointer that is distinct from the pointer that the GlobalThreatFeedCacher is
holding internally.

- The UT code (MockGlobalThreatFeedCache) does not have any form of
synchronization, and does not use `DeepCopy()`, so it returns the same pointer
to its callers as it is holding itself internally.

In the UT case, I don't think there can be a problem with read/write of the
pointer value, as this is never changing.  I think the problem must lie with
the read/write of the `<GlobalThreadFeed>.Status.ErrorConditions` slice value;
this is where we can be introducing undefined behaviour.

To fix that - and without requiring the callers to use a fine-grained mutex
when they modify parts of the pointer that they are given - we need to add
`DeepCopy()` to the MockGlobalThreatFeedCache, same as the production
equivalent already has.  But that means that the `cachedFeed` pointer value is
changing, so then we also need to add a mutex around the reading and writing of
that pointer value.