Fix SNI server name extraction for TLS 1.3 compatability with the crypto library
To get the servername that voltron needs to route reqeusts from a
managed cluster (i.e. es gateway vs linseed vs ia) we use the crypto tls
library to start the tls handshake process with the request sent from
the managed cluster. This is done so that the client hello can be parsed
from the request (which contains the servername). We expect the
handshake to fail because we initiate the handshake with a read only
connection, so when the crypto library attempts to write the server
hello it fails (but at this point it has already written the client
hello to the connection state it returns). So the process was:
- Parse the client hello
- Do some validation
- Store the client hello in the connection state returned
- Write the server hello (failed for a read only connection)

With the introduction of tls 1.3, the crypto library has changed the
code flow for tls 1.3 requests (it's still the previous flow for tls 1.2
requests). The process is now:
- Parse the client hello
- Do some validation
- Write the server hello (fails for a read only connection)
- Store the client hello in the connection state returned

So now the handshake fails before the client hello is stored in the
connection object.

This process probably wasn't the best way to do this anyway, given that
we had to also add bogus certificates to get past the initial validation
after the client hello was parsed to get to the point where it was
written to the connection object.

We will now take advantage of the GetConfigForClient hook, which, `if not
nil, is called after a ClientHello is received from a client`. So we use
this function to grab the client hello object right after it's parsed,
then return a custom error that we can check for and ignore to stop tls
processing.