Load ui / tunnel proxy targets from mounted files
This commit adds the ability to load proxy targets for both the UI and
the tunnel via a mounted file. This work is additive, so the hardcoded
routes are still there and you can add additional routes through the
file based approach.

This allows for adding both terminated tls and passthrough tls routes to
the tunnel, and only tls terminated routes to the UI.