[RS-1744] Implement /event-exceptions frontent API (#7235)
This PR implements the /event-exceptions frontend API described in https://docs.google.com/document/d/1HLaQuncsHt26G3R2-88zYYMCXGDTA9BjgXPvelu-_68/edit#heading=h.6xi0wn7olnwo.

In a nutshell, it exposes AlertExceptions as EventExceptions to the UI, to keep storing the rules as AlertExceptions while keeping some of the translation and business logic in the backend.

One goal of the epic is to facilitate the creation of exceptions for security events. User will create an exception from a false positive they want to ignore. We keep it simple initially.