fix svcgrp selector when it contains networkset
service group selector is built by having the service selector and adding the
selector from all the different types of endpoints that can be behind the service group.

networkset can also be aggregated in a service group but in building the selectors within
the servicegroup, we were missing to add networksets' selector.

This change makes sure GetServiceNodeSelectors and GetServicePortNodeSelectors will
include networksets' selector as well if it exists in the flow endpoint.

for all the other types that we do not indluce in the selector, we added a debug log.
So in future, if a service group filter is missing something, debugging can help
identify which unexpected type is aggregated behind the service group and we can decide
what to do for that case if it raises.

unit tests are added to specifically check for networkset selector to be included in
the service, servicegroup, and serviceport selectors.