Allow proxy target host header to be overwritten
For cases where we proxy to an ingres that proxies based off the host
name header, we need to be able to overwrite the host header. Currently
none of the services we proxy to do this, but there are some future
cases where this is an issue (proxying to ops cluster services).