update qs /policies to sort by more sorting and filtering options
Change list includes:
- updating policy resource cached in policyCache to include policy order field
- Adding "order" and "creation_time" to list of sorting parameters
- update tier in the request params to accept more than one tier, if tier is provided policy
form the tiers will be returned and if it is not specified it will search through all tiers.

No tests are added for the new functionalities as part of this change list since they cannot be
tested on their own and there is no existing test suits for the functions involved. The work is
tracked at #EV-4866