Patch and build envoy from source
This change patches and builds envoy from source. It creates an
intermediate `tigera/envoybinary` multi-arch image which holds our
customized envoy binary. No dependency is copied so it is not intended
to be deployed directly into a Calico Enterprise cluster. The final
`tigera/envoy` image will be reconstructed and published as part of
l7-collector build process.