Set $SEMAPHORE_GIT_COMMIT_RANGE to branch_range
This change sets `$SEMAPHORE_GIT_COMMIT_RANGE` to the `branch_range`
parameter for `change_in` in semaphore When condition. For a non-master
branch, semaphore will examine all commits between the head of the
current branch and the common ancestor for that branch [1]. This
behavior triggers many unnecessary rebuilds in our release branches.
After this change, the release branch behavior will be the same as the
master branch and only commits in the push will be examined.

[1] https://docs.semaphoreci.com/reference/conditions-reference/#change_in