dual-tor test: Use BGP layout instead of manual AS number annotation
Our dual ToR docs advise creating a bgp-layout ConfigMap to configure nodes' AS
numbers, but the dual ToR test was still using an earlier approach of manual
per-node annotation.  We now have a customer who is following the docs and
hitting a problem, so it's important we update our testing to align with that
and help repro the problem.

(The dual ToR development happened in two phases, with the bgp-layout idea
being part of the second phase, and we forgot to update this test to use it.)