Add retry limit and lots of diags for a failing dual ToR test
We had a problem where calico-nodes were not becoming ready in the dual ToR
test.  (Which was because they were failing to establish correct routing
between the nodes - but the detail of why is not the main point of _this_
commit.)

In that situation, dualtor.sh keeps retrying the readiness test forever - so
the only backstop is Semaphore's global time limit, and we don't get any
artifacts from the test when it runs on Semaphore.

Here we add a retry limit and also lots of logging and diags that should be
useful in debugging if this happens again.