Fail FV tests on core dumps and races.
We've had detection for races and core dumps for a long
time but unless we fail the tests, we don't get alerted
and we don't get the logs.

Detect core files/races at tear down and make sure the
error highlights that the main test has passed/failed
so as to avoid confusion.