Add user info to AuthorizationReview CRD (#7818)
Currently there is no way to pass user info via AuthorizationReview.Spec and we need to
pass in user via the request context. However this approach requires the app to have
"impersonate" rbac which is too much power to be given to any app that is just trying to
check for user's permissions.

Thus, introducing this change to allow apps to pass in user's info as part of the
AuthorizationReview instead of the request context.

This change also includes a fix in queryserver to use this new fields in AuthorizationReview
to get logged-in user's permissions on (tiered/ non-tiered) resources.