dual-tor test: Use current node code
dual-tor CI, for a PR or a branch other than master, was not using the actual
node code for that PR or branch.  Instead it was using the
gcr.io/unique-caldron-775/cnx/tigera/cnx-node:master image, built from the
latest calico-private master code, by way of using the master Tigera operator.

That's clearly no good when one is working on a PR to fix something in the node
code for dual ToR!

We can't easily tell the operator to use a different cnx-node image, so instead
we retag our locally built image to pretend that it is
gcr.io/unique-caldron-775/cnx/tigera/cnx-node:master.