Remove queryserver authz from es-proxy endpoint
calling authz to queryserver from es-proxy is calling issues for cloud and mcm
becasue es-proxy is in management and queryserver is managed. Thus, for user
with access to management clusters only, the autz from es-proxy is failing
since it is trying to authorize the user on the queryserver within the
same cluster as es-proxy (which is management cluster)

To solve the problem, we are passing logged-in user's token to queryserver
(for the call from es-proxy) and that includes logged-in user's token. then
queryserver can do it's own authorization checks.