Read endpoint from NonClusterHost resource in fluent-bit linseed plugin
This changeset watches NonClusterHost resource in the cluster for log
ingestion endpoint changes. When endpoint isn't specified in the
environment variable nor plugin configuration, the linseed out plugin
will read endpoint from the NonClusterHost resource and watch for
changes. This behavior should be the common use case for our customers.