Add retry loop for NonClusterHost resource get
This changeset adds an exponential retry loop over NonClusterHost
resource get. During the linseed out plugin initialization, it is possible
to fail when there are network or apiserver throttling issues. When
the error propagates back to systemd, it will restart the service
prematurely. This change also bumps fluent-bit version to v3.1.8.