Skip protobuf builds in CI
When arm64 builds are run in semaphore, Make often attributes an outdated
timestamp to the generated protobuf file, and thus triggers generation of
the protobuf file. Since generation is done on an arm64 machine, it
requires an arm64 protoc image, which does not exist.

A previous attempt to fix this issue tried to update the timestamp of
the generated protobuf file, but it appears to not be working
consistently.

This commit implements a skip mechanism for protobuf generation. This
skip is safe so long as the pre-flight checks for the component being
built ensure that the checked-in generated files are up to date. The
pre-flight checks are done on amd64, which has a protoc image.