Remove RESTConfig from controller configs (#8187)
When a RESTConfig gets recomputed, it's value may change as a result of
the mounted token being updated by the Kubernetes control plane. Since
multiple controllers have a RESTConfig in their configuration, when
configuration was recomputed after a token refresh, kube-controllers
would detect a configuration change and restart.

This restart is not necessary, since client-go has a mechanism to
refresh its own token when the mounted token is refreshed.

This change removes RESTConfig from controller configurations so that
restarts do not occur when tokens are refreshed. A kubeconfig path is
passed instead.