[EV-5358] Fix flowlog policymatch for k8s policies (#8188)
* cymatch for k8s policies

kubernetes policies name are generated differently from other
policies. Thus, we added a new field "type" to PolicyMatch to
be able to construct the string match for kubernetes policies properly.

Fix flowlog policymatch to match for k8s and staged policies

"kubernetes policies" and "staged policies" name are generated
differently from other policies. Thus, we need to add fields "type" and "staged"
to PolicyMatch to be able to construct the string match for kubernetes and
staged policies properly.

policy combined-name in flowlogs is constructed differently depending on the type of hit.
 - Namespaced policy combined-name: <namespace>/<tier>.<name>
 - Global / Profile combined-name: <tier>.<name>
 - namespaced policies: <namespace>/staged:<tier>.<name>
 - global policies: staged:<tier>.<name>
PolicyMatch.Staged defines if the "staged:" tag needs to be added to the policy name string match

* fix tests

* add testcases

* fix staged tag:should be added before the tier

* introduce policymatch.global for global policy search

we can have 3 types of policy match when it comes to namespaces:
- match for a specific namespace: ns value should be provided.
- match for any namespace: namespace value should not be provided.
- match for global policies: global flag needs to be set to true.

the default when neither the "global", nor the "namespace" values
are provided is: matching for namespaced-policies in any namespace.

* fix more tests