Increase Linseed server's WriteTimeout to 60s to fix TLS errors
Linseed has been returning "tls: bad record MAC" errors to its clients when a
request takes more than 10s to complete.  This doesn't necessarily depend on
the apparent size or complexity of the request.  For example, we occasionally
see "tls: bad record MAC" when Sasha tries to write a single Event to Linseed.
Presumably the problem occurs when the write triggers e.g. reindexing in
Elastic, or coincides with other Elastic processing that is already under way.

On its downstream side - i.e. acting as a client to the Elastic DB server -
Linseed specifies a timeout (`DefaultTimeOut`) of 60 seconds, and in some
cases (only) allows this to be overridden by a `Timeout` field in the request
params.  But on its upstream side - i.e. acting as a server for Linseed clients
- Linseed hard coded a timeout (`defaultWriteTimeout`) of 10 seconds.  This
change aligns the upstream timeout with the downstream one.

Many thanks to @gcosgrave for [debugging this](https://tigera.slack.com/archives/GKTBUHGN4/p1734534619145729?thread_ts=1712568963.512339&cid=GKTBUHGN4).