[RS-2324] Retry creating DPI security event when linseed is down (#8594)
## Description

When upgrading calico enterprise, there is a bug where DPI stops working if it fails to write an event to linseed in its startup phase. This can happen if linseed is down (because it is also restarting...).

This PR addresses the issue by retrying to write the DPI security event until it is successful.

## Related issues/PRs

https://tigera.atlassian.net/browse/RS-2324