Retain Service Graph edges for host-type nodes
This changeset adds a `retainEdges` flag to the Service Graph internal
`trackedNode`. For host-type SG nodes, they are tracked under the
same internal `trackedGroup`. When the SG graph constructor converts
raw l3 flow data into SG nodes and edges, it won't add an edge between
them. This logic is useful for nodes like service group (svcgp). However,
For (non-cluster) hosts, it seems to be a bug when traffic is flowing
between them but no edges are shown in the Service Graph.