Ignore output of failed DNS parser load in iptables
At some older system, it is expected that the DNS parser does not load
and we fall back to DelayDenied packet mode. The output is log and
disruptive.

We need to provide our own buffer to libbpf so that libbpf outputs in
that buffer and does not print itself. We ignore the output of the
buffer then.

https://github.com/libbpf/libbpf/blob/d4a841a32b04d69194ab5bdac359a51938a206ce/src/libbpf.c#L7582-L7584

If we switch Felix output to severity Debug, we also print failed DNS
parser loads so there is still a way how to debug a DNS parser load
issue.