[TSLA-8647] Add oiler for Elastic migrations
Oiler is a tool used for Calico Cloud to migrate ES data regardless of the format. This component is not part of Enterprise, but it will be build by our CI pipelines.

Its immediate application will be to migrate data from internal es single tenant to external es single tenant.

We will make use of Linseed's backend that can transform data in format (single-tenant/multi-tenant) and can connect with two different Elasticsearch clusters. The logic behind this migration is that we will perform a read from a primary source and write to secondary source.