fix kubernetes np uid
Kubernetes policies are converted to Calico policies and in the
conversion process, the UUID is converted. This converted UUID is
internal and conversion is happenign due to the internal storage.
Since we are returning the knp from queryserver API, we need to
return to actual UUID, not the internal UID.
Thus, this change is running the ConvertUID function for knp when
preparing the objects to be returned from the API.