[TSLA-8844] Allow IDs to be provided at ingestion time
Add migration mode to Linseed backends. Migration mode will require that:

    All list requests will be executed with a point in time ID and data returned will contain the Elasticsearch ID
    All create requests will preserve the provided Elasticsearch ID in order to migrate the same data across Elasticsearch IDs.

Elasticsearch ID will not be stored as document level, but exposed in the json in a migration mode.

Notes on Point in Time:

    A point in time is an Elastic API that creates a snapshot in time to perform deep pagination.
    If you provide a PIT ID, you must not pass index parameter to a seach request.
    Pagination is done via search_after parameters that are extracted from sort parameters from the response.
    A PIT request add by default a tie breaker ID, without the need to specify a tie breaker parameter for sort

This PR also contains changes to Oiler and the migration that needs
adjusting to these changes and a promotion build.